할인 정책 Enum으로 이동

+++ src/sparta/front/FrontPage.java
@@ -3,5 +3,5 @@ package sparta.front;
 import sparta.enumtype.DiscountEvent;
 
 public interface FrontPage {
-    void display(long price, DiscountEvent event);
+    void display(int price, DiscountEvent event);
 }
